Fix encoders

+++ MsfsPlugin/MsfsPlugin/encoder/AileronTrimEncoder.cs
@@ -7,10 +7,8 @@
     class AileronTrimEncoder : DefaultEncoder
     {
         public AileronTrimEncoder() : base("Aileron Trim", "Aileron trim encoder", "Misc", true, -100, 100, 1) { }
-
         protected override void RunCommand(String actionParameter) => this.SetValue(0);
         protected override Int32 GetValue() => MsfsData.Instance.CurrentAileronTrim;
-
-        protected override Int32 SetValue(Int32 newValue) => MsfsData.Instance.CurrentAileronTrim = (Int16)newValue;
+        protected override Int32 SetValue(Int32 newValue) => MsfsData.Instance.CurrentAileronTrim = newValue;
     }
 }
