+++ MsfsPlugin/MsfsPlugin/encoder/RudderTrimEncoder.cs
@@ -7,10 +7,8 @@
     class RudderTrimEncoder : DefaultEncoder
     {
         public RudderTrimEncoder() : base("Rudder Trim", "Rudder trim encoder", "Misc", true, -100, 100, 1) { }
-
         protected override void RunCommand(String actionParameter) => this.SetValue(0);
         protected override Int32 GetValue() => MsfsData.Instance.CurrentRudderTrim;
-
-        protected override Int32 SetValue(Int32 newValue) => MsfsData.Instance.CurrentRudderTrim = (Int16)newValue;
+        protected override Int32 SetValue(Int32 newValue) => MsfsData.Instance.CurrentRudderTrim = newValue;
     }
 }
